Uploading Code Journal 1 entry 2
Write a Python that prints the sum of two floating point numbers, the difference between two integers, and the product of a floating point number and an integer. In each case, have the program print out the data type of the resulting answer.

This program uses user input to print out the above.